[hyuk] <chore> conflict 해결 (#23)

+++ src/main/java/com/one/mycodi/service/PostService.java
@@ -140,12 +140,33 @@ public class PostService {
     }
 
     @Transactional
-    public ResponseDto<?> updatePost(Long id, PostRequestDto requestDto) {   // post 업데이트
+    public ResponseDto<?> updatePost(Long id, PostRequestDto requestDto,HttpServletRequest request) {   // post 업데이트
+        if (null == request.getHeader("Refresh-Token")) {
+            return ResponseDto.fail("MEMBER_NOT_FOUND",
+                    "로그인이 필요합니다.");
+        }
+
+        if (null == request.getHeader("Authorization")) {
+            return ResponseDto.fail("MEMBER_NOT_FOUND",
+                    "로그인이 필요합니다.");
+        }
+
+        Member member = validateMember(request);
+        if (null == member) {
+            return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
+        }
+
+
 
         Post post = isPresentPost(id);
         if (null == post) {
             return ResponseDto.fail("200", "존재하지 않는 게시글 id 입니다.");
         }
+
+        if (post.validateMember(member)) {
+            return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
+        }
+
         post.update(requestDto);
 
         return ResponseDto.success(
@@ -162,11 +183,32 @@ public class PostService {
     }
 
     @Transactional
-    public ResponseDto<?> deletePost(Long id) { // post 삭제
+    public ResponseDto<?> deletePost(Long id,HttpServletRequest request) { // post 삭제
+        if (null == request.getHeader("Refresh-Token")) {
+            return ResponseDto.fail("MEMBER_NOT_FOUND",
+                    "로그인이 필요합니다.");
+        }
+
+        if (null == request.getHeader("Authorization")) {
+            return ResponseDto.fail("MEMBER_NOT_FOUND",
+                    "로그인이 필요합니다.");
+        }
+
+        Member member = validateMember(request);
+        if (null == member) {
+            return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
+        }
+
+
         Post post = isPresentPost(id);
         if (null == post) {
             return ResponseDto.fail("200", "존재하지 않는 게시글 id 입니다.");
         }
+
+        if (post.validateMember(member)) {
+            return ResponseDto.fail("BAD_REQUEST", "작성자만 삭제할 수 있습니다.");
+        }
+
 
         postRepository.delete(post);
